rkvd/service.go: Insert future no longer returns an error
We panic if a client sends a requests with a clientID that has not been
registered. We do this because it's a programming error on our part if it
happens and it should be reported. Obviously this needs to be reverted for
actual production usage as we can't have clients crashing servers.